+++ Changelog.md
@@ -31,6 +31,7 @@ Rust:
 - update validity checks on memory management
 - intermingle gates and function declarations into a list of directives
 - rename Public/PrivateInput gates by Public/Private
+- replace table by struct for WireRange in Flatbuffers schema
 
 # SIEVE IR v1.0.1, 2022-08